requirements.txt
auto_mix_prep==0.2.0
libs==0.0.10
matplotlib==3.4.3
numpy==1.20.3
opencv_python==4.5.5.62
<<<<<<< HEAD
picamera==1.13
=======
>>>>>>> RL
Pillow==9.1.1
torch==1.11.0
